requirements.txt
boto3~=1.38.16
python-dotenv~=1.1.0
Flask~=3.1.1
flask-cors~=5.0.1
Werkzeug~=3.1.3
ollama~=0.4.8
pdfplumber~=0.11.6
langchain~=0.3.25
langchain-core~=0.3.60
langchain-ollama~=0.3.2
langchain-community~=0.3.24
